Do not automatically autodetect group members from looted items.
RELEASE NOTES:

The action of player looting an item no longer automatically adds it
to the group list. For example, when a player loots an Eternal item
in Gelkmaros or Ingission, the loot is broadcast to all players in
the area.
This caused APme to erroneously detect the player as being in the group.
Group members are still detected when rolling dice for items.