Implement the ?inv command. Also change the default invfull policy to OFF.
RELEASE NOTES:

Changed how the inventory full works. Players will not be excluded from
the AP loot system if their inventory is full. This was confusing for the
users. A warning is still shown. The old behavior can be enabled by using
the new "?inv on" command.

Implemented new ?inv command. It has 3 sub-commands ON/OFF/CLEAR:
    - ON: Enables inventory full restrictions (old behavior)
    - OFF: Disables inventory full restrictions (just warnings)
    - CLEAR: Disable the inventory full status

When inventory full restrictions are enabled, players with full inv
will be excluded from the AP loot sharing system.