Implement ?aplimit and loot rights limits.
RELEASE NOTES:

Implement the option to set AP upper limits in the loot system.

If a player exceeds the AP limit, it is excluded from loot rotation.
When all player reach the AP limit, the loot is free for all which
is also displayed in the loot rights.

This is commonly used to implement loot mechanics similar to the
RR2400 system.

The upper AP limit can be set with the ?aplimit command. When the
AP limit is set to 0, the upper limit is disabled.