RELEASE NOTES:
Add the ?apformat command. With this command you can change the format
of the ?aploot command. It takes 4 different arguments:

 - short: short format, shows only roll list
 - medium: medium format, shows roll and pass lists
 - long: long format, shows roll and pass lists with AP for each player
 - custom format

With the custom format a player can defined whatever the text looks like.

The custom format is specified by passing 6 formattingstrings, separated
with a "/":

/ROLL_HEADER/ROLL_LIST/PASS_HEADER/PASS_LIST/INV_HEADER/INV_LIST/

The _HEADER strings are used to prefix each list. While the _LIST strings
specify the format for each player in the list.

There are several keywords that you can use to customize the _LIST strings:
- @name will get replaced by players' name
- @ap will get replaced by the player's AP value
- @/ will get replaced by "/"

For example, the short list is basically just an alias for:
/ROLL(@ap):/ @name///INV FULL:/ @name/

You cannot use the @ keywoards in headers, except the ROLL header accepts @ap,
which will be replaced  by the lowest AP value in the group.

Formatting strings can be empty, in which case they wont be shown.

If the user wishes to customize the format permanently, a new environment
variable was added, APME_APFORMAT, which accepts the same format as the
?apformat command.